+++ osler/core/admin.py
@@ -3,10 +3,9 @@ from django.contrib import admin
 from osler.utils import admin as admin_utils
 from osler.core import models
 
-import osler.users.models as user_models
 
-for model in [user_models.Language, models.Patient,
-              user_models.Gender, models.ActionInstruction, models.Ethnicity,
+for model in [models.Language, models.Patient,
+              models.Gender, models.ActionInstruction, models.Ethnicity,
               models.ReferralType, models.ReferralLocation,
               models.ContactMethod, models.DocumentType, models.Outcome]:
     admin_utils.simplehistory_aware_register(model)
