+++ osler/core/models.py
@@ -14,7 +14,24 @@ from simple_history.models import HistoricalRecords
 from osler.core import validators
 from osler.core import utils
 
-from osler.users.models import Language, Gender
+
+class Language(models.Model):
+    """A natural language, spoken by a provider or patient.
+    """
+    name = models.CharField(max_length=50, primary_key=True)
+
+    def __str__(self):
+        return self.name
+
+
+class Gender(models.Model):
+    name = models.CharField(max_length=30, primary_key=True)
+
+    def __str__(self):
+        return self.name
+
+    def short_name(self):
+        return self.name[0]
 
 
 class ContactMethod(models.Model):
@@ -128,7 +145,7 @@ class Patient(Person):
     class Meta:
         permissions = [('can_case_manage_Patient', "Can act as a case manager.")]
 
-    case_managers = models.ManyToManyField(get_user_model())
+    case_managers = models.ManyToManyField(settings.AUTH_USER_MODEL)
 
     outcome = models.ForeignKey(Outcome, null=True, blank=True,
                                 on_delete=models.PROTECT)
@@ -384,7 +401,7 @@ class CompletableMixin(models.Model):
 
     completion_date = models.DateTimeField(blank=True, null=True)
     completion_author = models.ForeignKey(
-        get_user_model(),
+        settings.AUTH_USER_MODEL,
         blank=True, null=True,
         related_name="%(app_label)s_%(class)s_completed",
         on_delete=models.PROTECT)
