+++ osler/users/admin.py
@@ -12,6 +12,6 @@ class UserAdmin(auth_admin.UserAdmin):
 
     form = UserChangeForm
     add_form = UserCreationForm
-    fieldsets = (("User", {"fields": ("name",)}),) + auth_admin.UserAdmin.fieldsets
-    list_display = ["username", "name", "is_superuser"]
+    fieldsets = (("User", {"fields": ("name", "gender", "languages",)}),) + auth_admin.UserAdmin.fieldsets
+    list_display = ["username", "last_name", "first_name", "is_superuser"]
     search_fields = ["name"]
